% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Jprior_pldis.R
\name{Jprior_pldis}
\alias{Jprior_pldis}
\title{Jeffreys Prior for the Discrete Power-Law Model}
\usage{
Jprior_pldis(alpha, xmin)
}
\arguments{
\item{alpha}{A numeric value representing the scaling parameter \eqn{\alpha},
which must be greater than 1.}

\item{xmin}{A numeric value specifying the lower bound \eqn{x_{\min}}.}
}
\value{
A numeric value representing the Jeffreys prior evaluated at
the given scaling parameter.
}
\description{
Computes the Jeffreys prior for a discrete power-law model, which is a
non-informative prior commonly used in Bayesian inference to ensure parameter
invariance under reparameterization.
}
\details{
The Jeffreys prior for the discrete power-law model is derived from its
Fisher information. Its computation involves the second derivative of the
log-likelihood function with respect to \eqn{\alpha}, while keeping
\eqn{x_{\min}} fixed. It is given by:
\deqn{\pi_1\left(\alpha\right) \propto \sqrt{\frac{\zeta''(\alpha, x_{\min})}
{\zeta(\alpha, x_{\min})} - \left(\frac{\zeta'(\alpha, x_{\min})}
{\zeta(\alpha, x_{\min})}\right)^2}}

This implementation relies on the Hurwitz zeta function, provided by the
VGAM package, to accurately compute the required terms.
}
\references{
Clauset, A., Shalizi, C. R., & Newman, M. E. (2009). Power-law distributions
in empirical data. \emph{SIAM Review}, 51(4), 661-703.
}
