% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se_expected.R
\name{se_expected}
\alias{se_expected}
\title{Expected-Information Standard Error of the Scaling Exponent}
\usage{
se_expected(alpha, xmin, n)
}
\arguments{
\item{alpha}{A numeric value representing the scaling parameter \eqn{\alpha},
which must be greater than 1.}

\item{xmin}{A numeric value specifying the lower bound \eqn{x_{\min}}.}

\item{n}{A numeric value specifying the sample size.}
}
\value{
A numeric value representing the information standard
error of the scaling exponent.
}
\description{
Computes the asymptotic standard error of an estimator of the scaling
exponent, obtained from the Fisher information of the discrete
power-law model.
}
\details{
Let \eqn{I(\alpha)} denote the sample Fisher information of the scaling
exponent for the discrete power-law model, given by:
\deqn{I\left(\alpha\right) = n\left(\frac{\zeta''(\alpha, x_{\min})}
{\zeta(\alpha, x_{\min})} - \left(\frac{\zeta'(\alpha, x_{\min})}
{\zeta(\alpha, x_{\min})}\right)^2\right)}
}
\examples{
se_expected(2.0, 1, 100)
se_expected(2.5, 1, 100)
se_expected(2.0, 2, 250)

}
\references{
Jerez-Lillo, N., Rodrigues, F. A., Ferreira, P. H., & Ramos, P. L. (2025).
Bias Reduction and Goodness-of-Fit in Discrete Power-Law Models.
Submitted for publication.
}
\seealso{
\link{fisher_penalty_pldis} \link{se_observed}
}
