% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se_observed.R
\name{se_observed}
\alias{se_observed}
\title{Observed-Information Standard Error of the Scaling Exponent}
\usage{
se_observed(f, ahat)
}
\arguments{
\item{f}{A one-argument function of \eqn{\alpha} returning the
log-likelihood (or Fisher-penalized log-likelihood) to be differentiated,
e.g. \code{function(z) loglik_pldis(z, xmin, x)} or
\code{function(z) loglik_fd_pldis(z, xmin, x)}.}

\item{ahat}{A numeric value at which to evaluate the observed information,
typically the corresponding estimate of \eqn{\alpha} (e.g. the MLE or
\eqn{\hat\alpha_{Fd}}).}
}
\value{
A numeric value representing the observed-information standard
error of the scaling exponent.
}
\description{
Computes the standard error of an estimator of the scaling exponent from
the observed information.
}
\details{
The observed information is minus the second derivative of \code{f} at
\code{ahat}, computed numerically via \code{numDeriv::hessian}. The
resulting standard error is:
\deqn{\text{SE}(\hat\alpha) = \left[-f''(\hat\alpha)\right]^{-1/2}.}
}
\examples{
# Example data
x <- c(1, 1, 2, 2, 2, 2, 2, 3, 3, 3, 4, 4, 10)

ahat <- fit_pldis(x, xm = 1, penalized = FALSE)$alpha
se_observed(function(z) loglik_pldis(z, 1, x), ahat)

ahat_fd <- fit_pldis(x, xm = 1, penalized = TRUE)$alpha
se_observed(function(z) loglik_fd_pldis(z, 1, x), ahat_fd)

}
\references{
Jerez-Lillo, N., Rodrigues, F. A., Ferreira, P. H., & Ramos, P. L. (2025).
Bias Reduction and Goodness-of-Fit in Discrete Power-Law Models.
Submitted for publication.
}
\seealso{
\link{se_expected} \link{loglik_pldis} \link{loglik_fd_pldis}
}
