% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_pldis.R
\name{fit_xmin_pldis}
\alias{fit_xmin_pldis}
\alias{fit_pldis}
\title{Fitting the Discrete Power-Law Model}
\usage{
fit_xmin_pldis(x, bayesian = TRUE)

fit_pldis(x, xm = NULL, bayesian = TRUE)
}
\arguments{
\item{x}{A numerical vector of observed data.}

\item{bayesian}{A logical value indicating whether to use a Bayesian approach
to estimate \eqn{x_{\min}}. The default is \code{TRUE}. If \code{TRUE}, the function employs Bayesian estimation
to refine the maximum likelihood estimate. If \code{FALSE}, the frequentist method is used.}

\item{xm}{A numeric value specifying the lower bound \eqn{x_{\min}}.
The default is \code{NULL}. If not specified, the function will automatically
estimate \eqn{x_{\min}} from the data.}
}
\value{
\itemize{
\item \code{fit_xmin_pldis()}: Returns the estimated lower bound based on the data.
\item \code{fit_pldis()}: Returns a list containing both estimated parameters.
}
}
\description{
Estimates the parameters of a discrete power-law model, including the lower
bound \eqn{x_{\min}} and the scaling parameter \eqn{\alpha}, using a
frequentist or Bayesian approach.
}
\details{
This function estimates the parameters of a discrete power-law model in
two steps:

First, the lower bound \eqn{x_{\min}} is estimated using the Kolmogorov-Smirnov
(KS) statistic, defined as:
\deqn{K = \max |F_n(x) - F(x; \hat{\boldsymbol{\theta}})|.}

This method identifies the value of \eqn{x_{\min}} that best aligns the
empirical data distribution with the power-law model. Specifically, it selects
\eqn{x_{\min}} such that the cumulative distribution function (CDF) of the
observed data, restricted to values \eqn{x \geq x_{\min}}, closely matches the
CDF of the fitted power-law model over the same range (see Clauset et al., 2007
for details).

Second, the scaling parameter \eqn{\alpha} is estimated.
}
\examples{
# Example data
x <- c(1, 1, 2, 2, 2, 2, 2, 3, 3, 3, 4, 4, 10)

# Estimation of xmin:
# First, we estimate xmin using the frequentist approach (bayesian = F)
fit_xmin_pldis(x, bayesian = FALSE)
# Then, we estimate xmin using the Bayesian approach (bayesian = T)
fit_xmin_pldis(x, bayesian = TRUE)

# Estimation of the full discrete power-law model parameters:
# First, we fix xmin to 1 and use the frequentist approach to estimate
# alpha (bayesian = F)
fit_pldis(x, xm = 1, bayesian = FALSE)
# Next, we fix xmin to 1 and use the Bayesian approach to estimate
# both alpha and xmin (bayesian = T)
fit_pldis(x, xm = 1, bayesian = TRUE)

# If no xmin value is provided, the function automatically estimates it:
# Estimation of the full discrete power-law model with the frequentist approach
fit_pldis(x, bayesian = FALSE)
# Estimation with the Bayesian approach
fit_pldis(x, bayesian = TRUE)

}
\references{
Jerez-Lillo, N., Rodrigues, F. A., Ferreira, P. H., & Ramos, P. L. (2025).
Beyond the Power Law: Estimation, Goodness-of-Fit, and a Semiparametric
Extension in Complex Networks. arXiv preprint arXiv:2311.11200. Available at:
\url{https://arxiv.org/abs/2311.11200}
}
\seealso{
\link{post_pldis} \link{loglik_pldis}
}
