% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_pldis.R
\name{fit_xmin_pldis}
\alias{fit_xmin_pldis}
\alias{fit_pldis}
\title{Fitting the Discrete Power-Law Model}
\usage{
fit_xmin_pldis(x, bayesian = T)

fit_pldis(x, xm = NULL, bayesian = T)
}
\arguments{
\item{x}{A numeric vector of observed data values.}

\item{bayesian}{A logical value indicating whether to use a Bayesian approach
to estimate \eqn{x_{\min}}. If \code{TRUE}, the function employs Bayesian estimation
to refine the maximum likelihood estimate. If \code{FALSE}, the frequentist method is used.}

\item{xm}{A numeric value specifying the lower bound \eqn{x_{\min}}.
The default is NULL. If not specified, the function will automatically
estimate \eqn{x_{\min}} from the data.}
}
\value{
\itemize{
\item \code{fit_xmin_pldis()}: Returns a list containing the estimated lower bound
of the discrete power-law model based on the data.
\item \code{fit_pldis()}: Returns a list containing both estimated parameters.
}
}
\description{
Estimates the parameters of a discrete power-law model, including the lower
bound \eqn{x_{\min}} and the scaling parameter \eqn{\alpha}, using both
frequentist and Bayesian approaches.
}
\details{
This function estimates the lower bound \eqn{x_{\min}} of a discrete power-law
model using the Kolmogorov-Smirnov (KS) statistic, defined by:
\deqn{K = \max |F_n(x)-F(x; \hat{\boldsymbol{\theta}})|.}

The KS statistic compares the empirical cumulative distribution function
of the observed data with the theoretical cumulative distribution function
of the discrete power-law model.

This approach aims to find the value of \eqn{x_{\min}} that makes the
distribution of the data and the power-law model fit as closely as possible.
In other words, it chooses \eqn{x_{\min}} so that the cumulative distribution
function of the observed data, restricted to values greater than or equal to
\eqn{x_{\min}}, is as similar as possible to the CDF of the power-law model
fitted to the same range of data (see Clauset et. al. 2007 for details).

When \code{bayesian = TRUE}, a Bayesian framework is applied to estimate both
\eqn{x_{\min}} and the scaling parameter \eqn{\alpha}.
}
\examples{
# Example data
x <- c(1, 1, 2, 2, 2, 2, 2, 3, 3, 3, 4, 4, 10)

# Estimation of xmin using the fit_xmin_pldis function:
# First, we estimate xmin using the frequentist approach (bayesian = F)
fit_xmin_pldis(x, bayesian = F)
# Then, we estimate xmin using the Bayesian approach (bayesian = T)
fit_xmin_pldis(x, bayesian = T)

# Estimation of the full discrete power-law model parameters using fit_pldis:
# First, we fix xmin to 1 and use the frequentist approach to estimate
# alpha (bayesian = F)
fit_pldis(x, xm = 1, bayesian = F)
# Next, we fix xmin to 1 and use the Bayesian approach to estimate
# both alpha and xmin (bayesian = T)
fit_pldis(x, xm = 1, bayesian = T)

# If no xmin value is provided, the function will automatically estimate xmin:
# Estimation of the full discrete power-law model with the frequentist approach
fit_pldis(x, bayesian = F)
# Estimation with the Bayesian approach
fit_pldis(x, bayesian = T)

}
\references{
Jerez-Lillo, N., Rodrigues, F. A., Ferreira, P. H., & Ramos, P. L. (2025).
Beyond the Power Law: Estimation, Goodness-of-Fit, and a Semiparametric
Extension in Complex Networks. arXiv preprint arXiv:2311.11200. Available at:
\url{https://arxiv.org/abs/2311.11200}
}
\seealso{
\link{post_pldis} \link{loglik_pldis}
}
